updat specs on m2-ssd & enclosure for ssd-updates
specs on m2-ssd and enclosure used for ssd-updates

In 2020, the enclosure and transcend-m2-ssd shown in pdf was used to replaced the mechanical hard-drive in 3 laptops. A more conventional 2.5-inch sata SSD had to be used in the 4th and oldest laptop as it would not boot from the sabrent-enclosed ssd (tho the newer laptops with sabrent-enclosure boot fine).